[AUTO BATCH PLUGIN] Add sync_infer_request.cpp, seprate the sync request related operations from auto_batch.cpp
